feat: Implement a comprehensive request management system including backend models, API endpoints, and corresponding frontend pages and components.

+++ backend/app/models.py
@@ -16,7 +16,7 @@ class RequestStatus(str, Enum):
 
 class RequestItemBase(BaseModel):
     name: str
-    quantity: float
+    quantity: int
     unit: str
     price: float
